requirements.txt
# ==============================================================================
# EnterpriseHub - Base Production Dependencies
# ==============================================================================
# All base/production dependencies required for the core application.
# For development/testing extras, see requirements-dev.txt
# For ML-specific dependencies, see requirements-ml.txt
# ==============================================================================

# ==============================================================================
# Core Framework
# ==============================================================================
streamlit>=1.41.1
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# ==============================================================================
# Data Processing & Analysis
# ==============================================================================
pandas>=2.1.3
numpy>=1.26.2
scipy>=1.11.4
scikit-learn==1.4.0
joblib==1.3.2

# ==============================================================================
# Async I/O Performance
# ==============================================================================
aiofiles>=24.1.0
nest-asyncio>=1.5.8

# ==============================================================================
# Data Visualization
# ==============================================================================
plotly==5.18.0
geopy>=2.4.0

# ==============================================================================
# Market Data & Technical Analysis
# ==============================================================================
yfinance==0.2.33
ta==0.11.0

# ==============================================================================
# AI & NLP (lightweight — heavy ML deps use lazy imports)
# ==============================================================================
anthropic>=0.40.0
openai>=1.50.0
tiktoken>=0.7.0
google-generativeai>=0.8.0
textblob==0.17.1
langgraph==1.0.6
langchain-core>=0.1.0
langchain-anthropic>=0.1.0
langsmith>=0.1.0
fastmcp>=0.2.0
# NOTE: sentence-transformers, transformers, spacy removed — they pull PyTorch
# (~2GB) and cause Railway builds to timeout. Services use lazy imports.

# ==============================================================================
# GHL Integration & Backend Services
# ==============================================================================
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx[http2]>=0.26.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
apscheduler>=3.10.0
bleach>=6.0.0
twilio>=9.0.0
sendgrid>=6.11.0
SpeechRecognition>=3.10.0

# ==============================================================================
# Utilities
# ==============================================================================
python-dotenv==1.0.0
openpyxl==3.1.2
redis>=5.0.0
aioredis>=2.0.1
PyYAML>=6.0
reportlab>=4.0
qrcode[pil]>=7.4.2

# ==============================================================================
# Database & Production Dependencies
# ==============================================================================
psycopg2-binary>=2.9.0
asyncpg>=0.31.0
pgvector>=0.2.0
sqlalchemy>=2.0.0
aiohttp==3.13.3

# ==============================================================================
# Authentication & Security
# ==============================================================================
PyJWT>=2.8.0
aiosqlite>=0.19.0
alembic>=1.13.0
cryptography>=41.0.0

# ==============================================================================
# Real-time Communication & System Monitoring
# ==============================================================================
websockets>=12.0
python-socketio>=5.11.0
python-engineio>=4.8.0
psutil>=5.9.0

# ==============================================================================
# Billing Integration
# ==============================================================================
stripe>=8.0.0

# ==============================================================================
# High-Performance Computing (from requirements-production.txt)
# ==============================================================================
aioboto3>=12.3.0
boto3>=1.34.0
botocore>=1.34.0
cachetools>=5.3.2
cloudpickle>=3.0.0
hiredis>=2.2.3
httptools>=0.6.1
msgpack>=1.0.7
orjson>=3.9.10
prometheus-client>=0.19.0
structlog>=23.2.0
uvloop>=0.19.0
